Created Data and Component for clarity
Trying to make the code self documenting. Clearer for Data access to be
in Data.js and Component rendering to be in Component.js. Don't have to
understand the app to know what is in each file
